% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotOn.R
\name{plotOn}
\alias{plotOn}
\title{Geographical representation of a data frame or raster map.}
\usage{
plotOn(x, map = ggmap::get_map(location = "trentino", zoom = 9),
  latlon_crs = as.character("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"),
  layer = 1, legend = "bottomright", title = "Soil Water Content",
  label = "swc", high = "blue", low = "white", alpha = c(0.1, 0.8),
  facet_wrap = FALSE, nrow = NULL, ncol = NULL, range = NULL,
  scale = "continuous_2", scale.fill.gradient = TRUE, plot = TRUE,
  scale_colour_gradientn_scale = NULL, extent = "normal",
  method = c("bilinear", "ngb"), palette = rainbow, shape = 15,
  size = NA, ...)
}
\arguments{
\item{x}{a \code{Raster*} class object}

\item{map}{geograghical map on which \code{x} is plotted. It is an object returned by \code{\link{get_map}} or similars.}

\item{latlon_crs}{string containing the utilized latitude longitude Coordinate Refarance System. See default in \code{Usage}.}

\item{layer}{brick layer utilized for geographical plotting. Default is 1.}

\item{title}{string title of the graphic}

\item{label}{string title (label) of the legend. It is used as the name of the scale if \code{scale.fill.gradient} is \code{TRUE}, otherwise it is ignored.}

\item{high}{colour for low end of gradient. See \code{\link{scale_fill_gradient}}.}

\item{low}{colourf or high end of gradient. See \code{\link{scale_fill_gradient}}.}

\item{alpha}{alpha coefficient. See \url{http://en.wikipedia.org/wiki/Alpha_compositing}.}

\item{facet_wrap}{logical value. If \code{TRUE} it uses \code{\link{facet_wrap}} to print all plots.}

\item{nrow,ncol}{number of rows and columns. See \code{\link{facet_wrap}}.}

\item{scale}{alternative parameter to \code{scale.fill.gradient}, it is a term potentially added for color scale (see \code{\link{scale_colour_hue}} o similars). It is used only if it is not \code{NULL} or \code{scale.fill.gradient} is \code{FALSE}}

\item{scale.fill.gradient}{logical value. If it is \code{TRUE} (Default), it uses \code{\link{scale.fill.gradient}}}

\item{plot}{logical value. If \code{TRUE} (Default) the function also plots the map, otherwise returns its value without any preliminary plot.}

\item{scale_colour_gradientn_scale}{color scale for \code{\link{scale_colour_gradientn}}. It is used if it is nott \code{NULL}. Default is \code{NULL}.}

\item{method}{see \code{\link{projectRaster}}}

\item{palette}{palette function used for \code{\link{scale_fill_manual}} .}

\item{shape,size}{see \code{\link{geom_point}}}

\item{...}{further arguments}
}
\value{
A \code{"ggplot"} object.
}
\description{
Geographical representation of a data frame or raster map.
}
\note{
Useful link: \url{http://journal.r-project.org/archive/2013-1/kahle-wickham.pdf}
}
\examples{
library(geotopVis)
data(soilwatercontent)

x <- plotOn(x=soilwatercontent[[1]])

##' ANOTHER EXAMPLE
library(geotopbricks)
# The data containing in the link are only for educational use
wpath <- "http://www.rendena100.eu/public/geotopbricks/simulations/idroclim_test1"
tz <-  "Etc/GMT-1"
when <- as.POSIXct("2002-03-22",tz=tz)

# a 2D map:
x_e <- "SnowDepthMapFile"
# Not Run: uncomment the following line
m <- rasterFromOutput2DMap(x_e,when=when,wpath=wpath,timestep="OutputSnowMaps",
                           tz=tz,use.read.raster.from.url=TRUE)
## NOTE: set use.read.raster.from.url=FALSE (default)
# if the "wpath" directorty is in the local file system.
# Not Run: uncomment the following line
m/10

plot(m/10,main="Snow Depth")



 location = c(10.5,45.5,12.5,46.5)
 map=get_map(location = location) ###, zoom = 9)


xmap <- plotOn(m/10,map=map,high="blue",low="white",title="Snow Depth",label="depth[cm]",plot=TRUE,alpha=c(0,0.3),extent="device")
xmap
}
\seealso{
\code{\link{geom_point}},\code{\link{ggmap}},\code{\link{facet_wrap}}
}

